+++ src/main/java/frc/robot/commands/ArmHold.java
@@ -1,7 +1,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.Arm.Arm;
+import frc.robot.subsystems.arm.Arm;
 
 public class ArmHold extends Command {
     private Arm arm;
