+++ src/main/java/frc/robot/subsystems/Arm/Arm.java
@@ -2,14 +2,14 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.Arm;
+package frc.robot.subsystems.arm;
 
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Constants.ArmConstants;
 import frc.robot.Constants.SimulationConstants;
-import frc.robot.subsystems.Elevator.MechVisualizer;
+import frc.robot.subsystems.elevator.MechVisualizer;
 import frc.robot.util.SmarterDashboard;
 import org.littletonrobotics.junction.Logger;
