+++ src/main/java/frc/robot/subsystems/Elevator/ElevatorIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.Elevator;
+package frc.robot.subsystems.elevator;
 
 import org.littletonrobotics.junction.AutoLog;
