+++ src/main/java/frc/robot/subsystems/EndEffector/EndEffectorIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.EndEffector;
+package frc.robot.subsystems.endeffector;
 
 import org.littletonrobotics.junction.AutoLog;
